requirements.txt
astroid==3.3.5
blinker==1.9.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
docopt==0.6.2
filelock==3.16.1
Flask==3.1.0
identify==2.6.3
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
PyJWT==2.10.1
pylint==3.3.2
pytest==8.3.4
pytest-watch==4.2.0
PyYAML==6.0.2
tomlkit==0.13.2
virtualenv==20.28.0
watchdog==6.0.0
Werkzeug==3.1.3
